requirements.txt
streamlit
langchain
langchain_groq
langgraph
langchain_community
huggingface_hub
duckduckgo-search
python-dotenv
arxiv
wikipedia